# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posture/sessions/start": {
      "post": {
        "description": "Returns a session with an `id`. Store this `id` to call `sessions/end` and `data/batch` later.",
        "operationId": "PostureController_startSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session started",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostureSessionDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start a new posture session",
        "tags": [
          "Posture"
        ]
      }
    },
    "/api/posture/sessions/end": {
      "post": {
        "description": "Closes the session by setting `endTime` and summary stats. Call this when the user stops wearing the device.",
        "operationId": "PostureController_endSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session ended",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostureSessionDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "End an active posture session",
        "tags": [
          "Posture"
        ]
      }
    },
    "/api/posture/data/batch": {
      "post": {
        "description": "Send multiple sensor readings at once. Recommended: batch every 60 readings (~1 min of data at 1 Hz). Returns count of rows saved.",
        "operationId": "PostureController_batchSensorData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSensorDataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sensor data saved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/BatchInsertResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload sensor data in batch",
        "tags": [
          "Posture"
        ]
      }
    },
    "/api/posture/calibrate": {
      "post": {
        "description": "**No request body needed.** Call this while the user is sitting/standing in their natural upright position. The returned reference axis values define what \"upright\" means for this user. Store these values on the device to use as the baseline for future posture detection.",
        "operationId": "PostureController_calibrate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Calibration successful",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/CalibrationResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Calibrate to neutral posture",
        "tags": [
          "Posture"
        ]
      }
    },
    "/api/posture/sessions": {
      "get": {
        "description": "Returns sessions newest-first. Use `page` and `limit` query params to paginate.",
        "operationId": "PostureController_getSessions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PaginatedSessionsResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get paginated posture sessions",
        "tags": [
          "Posture"
        ]
      }
    },
    "/api/posture/sessions/{id}/stats": {
      "get": {
        "description": "Returns the session + computed stats (upright %, score, duration).",
        "operationId": "PostureController_getSessionStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stats retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostureSessionStatsResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get detailed stats for a session",
        "tags": [
          "Posture"
        ]
      }
    },
    "/api/posture/score": {
      "get": {
        "description": "Averages `averageScore` across all completed sessions today. Returns `score: null` with a message if no sessions completed yet today.",
        "operationId": "PostureController_getScore",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Score retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostureScoreResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get today's posture score",
        "tags": [
          "Posture"
        ]
      }
    },
    "/api/goals/daily": {
      "get": {
        "description": "Returns the goals set for today. If no custom goals are set, returns system defaults (30 min exercise, 8h upright, 1h device). Check `isDefault: true` for this case.",
        "operationId": "GoalsController_getDailyGoal",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daily goals retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DailyGoalResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get today's daily goals",
        "tags": [
          "Goals"
        ]
      },
      "put": {
        "description": "Creates or replaces the goals for today. Send only the fields you want to change.",
        "operationId": "GoalsController_updateDailyGoal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDailyGoalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Daily goals updated",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DailyGoalResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set today's daily goals",
        "tags": [
          "Goals"
        ]
      }
    },
...
  },
  "info": {
    "title": "Straight API",
    "description": "Backend API for the Straight posture correction mobile app. All endpoints except /api/users/sync require a Firebase Bearer token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PostureSessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "userId": {
            "type": "string",
            "example": "uuid-v4"
          },
          "startTime": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T09:00:00.000Z"
          },
          "endTime": {
            "type": "object",
            "example": "2024-01-15T09:30:00.000Z"
          },
          "uprightTimeMinutes": {
            "type": "object",
            "example": 25.5,
            "description": "Minutes spent upright"
          },
          "totalTimeMinutes": {
            "type": "object",
            "example": 30,
            "description": "Total session duration in minutes"
          },
          "averageScore": {
            "type": "object",
            "example": 82.5,
            "description": "Score 0–100"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "startTime",
          "createdAt"
        ]
      },
      "SessionStatsDto": {
        "type": "object",
        "properties": {
          "totalDataPoints": {
            "type": "number",
            "example": 360
          },
          "uprightDataPoints": {
            "type": "number",
            "example": 300
          },
          "uprightPercent": {
            "type": "number",
            "example": 83,
            "description": "Percentage of time upright"
          },
          "averageScore": {
            "type": "object",
            "example": 82.5
          },
          "durationMinutes": {
            "type": "object",
            "example": 30
          }
        },
        "required": [
          "totalDataPoints",
          "uprightDataPoints",
          "uprightPercent"
        ]
      },
      "PostureSessionStatsResponseDto": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/PostureSessionDto"
          },
          "stats": {
            "$ref": "#/components/schemas/SessionStatsDto"
          }
        },
        "required": [
          "session",
          "stats"
        ]
      },
      "BatchInsertResponseDto": {
        "type": "object",
        "properties": {
          "inserted": {
            "type": "number",
            "example": 60,
            "description": "Number of sensor data rows saved"
          }
        },
        "required": [
          "inserted"
        ]
      },
      "CalibrationResponseDto": {
        "type": "object",
        "properties": {
          "calibrated": {
            "type": "boolean",
            "example": true
          },
          "referenceAxisX": {
            "type": "number",
            "example": 0,
            "description": "Reference X-axis value at neutral posture"
          },
          "referenceAxisY": {
            "type": "number",
            "example": 0,
            "description": "Reference Y-axis value at neutral posture"
          },
          "referenceAxisZ": {
            "type": "number",
            "example": 9.81,
            "description": "Reference Z-axis value at neutral posture"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "calibrated",
          "referenceAxisX",
          "referenceAxisY",
          "referenceAxisZ",
          "timestamp"
        ]
      },
      "PostureScoreResponseDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "object",
            "example": 78,
            "description": "Today's average score 0–100, null if no sessions yet"
          },
          "sessionsToday": {
            "type": "number",
            "example": 2
          },
          "totalUprightMinutes": {
            "type": "number",
            "example": 95.5,
            "description": "Total upright minutes today"
          },
          "message": {
            "type": "string",
            "example": "No completed sessions today"
          }
        }
      },
      "PaginatedSessionsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostureSessionDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 45
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "items",
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "StartSessionDto": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "example": "2024-01-15T09:00:00Z"
          },
          "deviceId": {
            "type": "string",
            "example": "device-uuid-here"
          }
        },
        "required": [
          "startTime"
        ]
      },
      "EndSessionDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "example": "session-uuid-here"
          },
          "endTime": {
            "type": "string",
            "example": "2024-01-15T09:30:00Z"
          },
          "uprightTimeMinutes": {
            "type": "number",
            "example": 25.5
          },
          "totalTimeMinutes": {
            "type": "number",
            "example": 30
          },
          "averageScore": {
            "type": "number",
            "example": 82.5
          }
        },
        "required": [
          "sessionId",
          "endTime"
        ]
      },
      "SensorDataItemDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "example": "2024-01-15T09:05:00Z"
          },
          "axisX": {
            "type": "number",
            "example": 0.12
          },
          "axisY": {
            "type": "number",
            "example": -0.05
          },
          "axisZ": {
            "type": "number",
            "example": 9.81
          },
          "isUpright": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "timestamp",
          "axisX",
          "axisY",
          "axisZ",
          "isUpright"
        ]
      },
      "BatchSensorDataDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "example": "session-uuid-here"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SensorDataItemDto"
            }
          }
        },
        "required": [
          "sessionId",
          "data"
        ]
      },
      "DailyGoalResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4",
            "description": "null if returning defaults"
          },
          "userId": {
            "type": "string",
            "example": "uuid-v4"
          },
          "exerciseDurationMinutes": {
            "type": "number",
            "example": 30,
            "description": "Exercise goal in minutes"
          },
          "uprightTimeMinutes": {
            "type": "number",
            "example": 480,
            "description": "Upright time goal in minutes (8 hours)"
          },
          "deviceUsageMinutes": {
            "type": "number",
            "example": 60,
            "description": "Device usage goal in minutes"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T00:00:00.000Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "isDefault": {
            "type": "boolean",
            "example": true,
            "description": "true when no custom goal set today — showing system defaults"
          }
        },
        "required": [
          "exerciseDurationMinutes",
          "uprightTimeMinutes",
          "deviceUsageMinutes",
          "date"
        ]
      },
      "UpdateDailyGoalDto": {
        "type": "object",
        "properties": {
          "exerciseDurationMinutes": {
            "type": "number",
            "example": 30,
            "description": "Exercise duration in minutes"
          },
          "uprightTimeMinutes": {
            "type": "number",
            "example": 480,
            "description": "Upright time goal in minutes"
          },
          "deviceUsageMinutes": {
            "type": "number",
            "example": 60,
            "description": "Device usage goal in minutes"
          }
        }
      },
...
    }
  }
}